+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/abstracts/types/ReplyMarkup.kt
@@ -1,7 +1,0 @@
-package dev.inmo.tgbotapi.abstracts.types
-
-import dev.inmo.tgbotapi.types.buttons.KeyboardMarkup
-
-interface ReplyMarkup {
-    val replyMarkup: KeyboardMarkup?
-}